Add tabs, this is still improvement, check it
