+++ Projects/AdventureCam/AdventureCamApp/AdventureCamApp.csproj
@@ -62,6 +62,7 @@
     <PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="$(MauiVersion)" />
     <PackageReference Include="Microsoft.Maui.Essentials" Version="$(MauiVersion)" />
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0" />
+    <PackageReference Include="Sharpnado.Tabs.Maui" Version="3.0.1" />
     <PackageReference Include="SimpleToolkit.Core" Version="4.0.0" />
   </ItemGroup>
 
@@ -71,6 +72,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\AdventureCamCore\AdventureCamCore.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <MauiXaml Update="Views\CameraPage.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </MauiXaml>
   </ItemGroup>
 
   <PropertyGroup Label="MauiAppAccelerator">
