+++ Projects/AdventureCam/AdventureCamApp/App.xaml
@@ -1,14 +1,23 @@
-﻿<?xml version = "1.0" encoding = "UTF-8" ?>
-<Application xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-			 xmlns:local="clr-namespace:AdventureCamApp"
-			 x:Class="AdventureCamApp.App">
-	<Application.Resources>
-		<ResourceDictionary>
-			<ResourceDictionary.MergedDictionaries>
-				<ResourceDictionary Source="Resources/Styles/Colors.xaml" />
-				<ResourceDictionary Source="Resources/Styles/Styles.xaml" />
-			</ResourceDictionary.MergedDictionaries>
-		</ResourceDictionary>
-	</Application.Resources>
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<Application
+    x:Class="AdventureCamApp.App"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:local="clr-namespace:AdventureCamApp">
+    <Application.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="Resources/Styles/Colors.xaml" />
+                <ResourceDictionary Source="Resources/Styles/Styles.xaml" />
+                <ResourceDictionary>
+                    <Style x:Key="TabStyle" TargetType="tabs:UnderlinedTabItem">
+                        <Setter Property="SelectedTabColor" Value="{StaticResource AccentColor}" />
+                        <Setter Property="FontFamily" Value="{StaticResource FontSemiBold}" />
+                        <Setter Property="LabelSize" Value="14" />
+                        <Setter Property="UnselectedLabelColor" Value="White" />
+                    </Style>
+                </ResourceDictionary>
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+    </Application.Resources>
 </Application>
