+++ Projects/AdventureCam/AdventureCamApp/AppShell.xaml
@@ -1,30 +1,9 @@
 ﻿<?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-	x:Class="AdventureCamApp.AppShell"
-	xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-	xmlns:views="clr-namespace:AdventureCamApp.Views"
-	xmlns:local="clr-namespace:AdventureCamApp">
-
-    <TabBar>
-        <ShellContent
-			Icon="iconblank.png"
-			ContentTemplate="{DataTemplate views:MainPage}" />
-        <ShellContent
-			Title="Preview"
-			Icon="iconmediaelement.png"
-			ContentTemplate="{DataTemplate views:PreviewPage}" />
-        <ShellContent
-			Title="Gallery"
-			Icon="iconblank.png"
-			ContentTemplate="{DataTemplate views:GalleryPage}" />
-        <ShellContent
-			Title="Editor"
-			Icon="iconlistdetail.png"
-			ContentTemplate="{DataTemplate views:EditorPage}" />
-        <ShellContent
-			Title="About"
-			Icon="iconblank.png"
-			ContentTemplate="{DataTemplate views:AboutPage}" />
-    </TabBar>
+    x:Class="AdventureCamApp.AppShell"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:local="clr-namespace:AdventureCamApp"
+    xmlns:views="clr-namespace:AdventureCamApp.Views">
+    <ShellContent ContentTemplate="{DataTemplate views:MainPage}" Icon="iconblank.png" />
 </Shell>
