+++ Projects/AdventureCam/AdventureCamApp/Views/CameraPage.xaml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage
+    x:Class="AdventureCamApp.Views.CameraPage"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:cv="clr-namespace:Camera.MAUI;assembly=Camera.MAUI"
+    Title="CameraPage">
+    <VerticalStackLayout>
+        <cv:CameraView x:Name="cameraView" HeightRequest="400" />
+    </VerticalStackLayout>
+</ContentPage>
