+++ Projects/AdventureCam/AdventureCamApp/Views/MainPage.xaml
@@ -3,10 +3,35 @@
     x:Class="AdventureCamApp.Views.MainPage"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:cv="clr-namespace:Camera.MAUI;assembly=Camera.MAUI"
+    xmlns:tabs="http://sharpnado.com"
     xmlns:vm="clr-namespace:AdventureCamApp.ViewModels"
     x:DataType="vm:MainViewModel">
-    <VerticalStackLayout>
-        <cv:CameraView x:Name="cameraView" HeightRequest="400" />
-    </VerticalStackLayout>
+    <!--<tabs:TabHostView
+        Grid.Row="2"
+        Margin="15"
+        BackgroundColor="#F0F0F3"
+        CornerRadius="30"
+        HeightRequest="60"
+        HorizontalOptions="Center"
+        IsSegmented="True"
+        SelectedIndex="{Binding Source={x:Reference Switcher}, Path=SelectedIndex, Mode=TwoWay}"
+        Shades="{sh:NeumorphismShades}"
+        TabType="Fixed"
+        VerticalOptions="Start"
+        WidthRequest="280">
+        <tabs:TabHostView.Tabs>
+            <tabs:BottomTabItem
+                IconImageSource="house_96"
+                Label="{localization:Translate Tabs_Home}"
+                Style="{StaticResource BottomTabStyle}" />
+            <tabs:BottomTabItem
+                IconImageSource="list_96"
+                Label="{localization:Translate Tabs_List}"
+                Style="{StaticResource BottomTabStyle}" />
+            <tabs:BottomTabItem
+                IconImageSource="grid_view_96"
+                Label="{localization:Translate Tabs_Grid}"
+                Style="{StaticResource BottomTabStyle}" />
+        </tabs:TabHostView.Tabs>
+    </tabs:TabHostView>-->
 </ContentPage>
